% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-align.R
\name{align-embedding}
\alias{align-embedding}
\alias{align_embedding}
\alias{rotate_embedding}
\title{Align/rotate one embedding onto another}
\usage{
align_embedding(x, y, ...)

rotate_embedding(x, y, ...)
}
\arguments{
\item{x}{a numeric matrix of unaligned embedding coordinates}

\item{y}{a numeric matrix to align `x`` to}

\item{...}{other arguments to pass \link[BiocSingular:runSVD]{BiocSingular::runSVD}}
}
\value{
A matrix of dimension equal to \code{x}
}
\description{
Align/rotate one embedding onto another
}
\details{
This is a (naive implementation) of Procuste's analysis,
the goal is to find a rotation matrix that most closely aligns the
matrix \code{x} to the matrix \code{y} by centering and rescaling. One method of solving this problem
is to perform singular value decomposition on
\code{crossprod(y - colMeans(y), x - colMeans(x))}.
}
