% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/TE-basisSets.R
\docType{methods}
\name{basisSets}
\alias{basisSets}
\title{Access or set the basisSets  of a TourExperiment}
\usage{
basisSet(x, type, ...)

basisSet(x, type, ...) <- value

basisSets(x, ...)

basisSets(x, ...) <- value

basisSetNames(x, type, ...)

basisSetNames(x, type, ...) <- value

\S4method{basisSets}{TourExperiment}(x)

\S4method{basisSet}{TourExperiment,missing}(x, type)

\S4method{basisSet}{TourExperiment,numeric}(x, type)

\S4method{basisSet}{TourExperiment,character}(x, type)

\S4method{basisSet}{TourExperiment,missing}(x, type, ...) <- value

\S4method{basisSet}{TourExperiment,numeric}(x, type, ...) <- value

\S4method{basisSet}{TourExperiment,character}(x, type, ...) <- value

\S4method{basisSetNames}{TourExperiment}(x)

\S4method{basisSetNames}{TourExperiment}(x) <- value
}
\arguments{
\item{x}{an object (a TourExperiment)}

\item{type}{the name or a number of the list to access}

\item{value}{a matrix to assign to an element of basisSets}
}
\value{
a \code{TourExperiment} object
}
\description{
Access or set the basisSets  of a TourExperiment
}
