% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/utils.R
\docType{methods}
\name{compute_half_range}
\alias{compute_half_range}
\alias{compute_half_range,ANY-method}
\alias{compute_half_range,LinearEmbeddingMatrix-method}
\title{Compute range of axes for bases}
\usage{
compute_half_range(.data)

\S4method{compute_half_range}{ANY}(.data)

\S4method{compute_half_range}{LinearEmbeddingMatrix}(.data)
}
\arguments{
\item{.data}{A 'matrix' like object}
}
\description{
Compute range of axes for bases
}
\details{
This function computes the maximum squared
Euclidean distance of rows in a matrix like object. Mostly used
internally for setting up xy-axis ranges for a tour animation.
}
