% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Preprocess a SightSE via dimension reduction}
\usage{
decompose(.data, num_comp, center, scale,
  .parallel = BiocParallel::SerialParam(), .method)
}
\arguments{
\item{.data}{A \code{SightSE}}

\item{num_comp}{Number of components to retain}

\item{center}{Should columns be centered?}

\item{scale}{Should  columns be scaled to unit variance?}

\item{.parallel}{A BiocParallel::BPPARAM object}

\item{.method}{A BiocSingular object}
}
\description{
Preprocess a SightSE via dimension reduction
}
