% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-linear.R
\name{embed_linear}
\alias{embed_linear}
\title{Compute a linear embedding on a \code{SightSE} object.}
\usage{
embed_linear(.data, num_comp, center = TRUE, scale = FALSE,
  .parallel = BiocParallel::SerialParam(), .engine)
}
\arguments{
\item{.data}{A \code{sneezy::SightSE()} object}

\item{num_comp}{Number of components to retain}

\item{center}{Should columns be centered? Default = TRUE}

\item{scale}{Should  columns be scaled to unit variance?}

\item{.parallel}{A BiocParallel::BPPARAM() object}

\item{.engine}{A BiocSingular::BiocSingular() object
used to compute the embedding. Deafults to \code{pca_exact()}}
}
\description{
Compute a linear embedding on a \code{SightSE} object.
}
