% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{embed_nonlinear}
\alias{embed_nonlinear}
\title{Compute a non-linear embedding on a \code{TourExperiment} object.}
\usage{
embed_nonlinear(.data, num_comp, .on = NULL, normalize = TRUE,
  .parallel = BiocParallel::SerialParam(), .engine = tsne_approx(30,
  12, 0.5))
}
\arguments{
\item{.data}{A \code{TourExperiment()} object}

\item{num_comp}{Number of components/dimensions to retain}

\item{.on}{Where to compute the non-linear embedding? If NULL,
the first assay is computed.}

\item{normalize}{Avoid numerical precision issues by centering and scaling the
input data. (Default = TRUE).}

\item{.parallel}{Register a parallel backend, only used if computing
nearest-neighbors prior to perfoming a non-linear embedding.}

\item{.engine}{A \code{NonLinearEmbeddingParam`` object used to define the  arguments and algorithm for non-linear embedding. Deafults to }tsne_approx()`}
}
\description{
Compute a non-linear embedding on a \code{TourExperiment} object.
}
\details{
This function computes a non-linear embedding over the parameters
for a non-linear embedding algorithm and adjusts the reducedDim slot with
the embedding coordinates.
}
\seealso{
\code{Rtsne::Rtsne()}
}
