% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{estimate_neighbors}
\alias{estimate_neighbors}
\title{Estimate k-nearest neighbours graph from a TourExperiment object}
\usage{
estimate_neighbors(.data, num_neighbors, .on = NULL,
  .engine = BiocNeighbors::KmknnParam())
}
\arguments{
\item{.data}{a \code{TourExperiment} object}

\item{num_neighbors}{An integer scalar for number of neighbors}

\item{.on}{a character scalar, indicating the part of \code{.data} to
estimate nearest neighbors. If missing the first assay will be used.}

\item{.engine}{A \code{BiocNeighbor::BiocNeighborParam()} object that
reperesents the algorithm used to compute nearest neighbors.}
}
\value{
A \code{TourExperiment} object with the \code{neighborSets} slot
filled.
}
\description{
Estimate k-nearest neighbours graph from a TourExperiment object
}
\details{
The representation of the nearest neighbors is a matrix of
size ncol(.data) by num_neighbors containing the neighbor indexes for
each sample in \code{.data}
}
