% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{generate_bases}
\alias{generate_bases}
\title{Generate a new basis set via a tour}
\usage{
generate_bases(.data, .on = NULL, clamp = FALSE, max_bases = 100,
  start = NULL, step_size = Inf, .engine = tourr::grand_tour())
}
\arguments{
\item{.data}{a TourExperiment object or matrix-like object}

\item{.on}{Which part of \code{.data} to tour}

\item{clamp}{Should the columns of toured data be clamped to lie in \link{0,1}?}

\item{max_bases}{Maximum number of bases to generate}

\item{start}{Optional starting projection for tour}

\item{step_size}{Distance between each step, set to \code{Inf} which forces new basis generation}

\item{.engine}{A tour path generator (defaults to \code{\link[tourr:grand_tour]{tourr::grand_tour()}})}
}
\description{
Generate a new basis set via a tour
}
