% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generate_sphere}
\alias{generate_sphere}
\title{Sample from a p-dimensional solid sphere}
\usage{
generate_sphere(n, p, mean, sd)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of dimensions}

\item{mean, sd}{passed to \link[stats:rnorm]{stats::rnorm}}
}
\description{
Sample from a p-dimensional solid sphere
}
\examples{
generate_sphere(1000, 10, mean =  5, sd = 2)
}
