% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles.R
\name{get_triangles}
\alias{get_triangles}
\title{Compute Delany triangulation in t-SNE space}
\usage{
get_triangles(coords)
}
\arguments{
\item{coords}{output from \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}}}
}
\description{
Compute Delany triangulation in t-SNE space
}
