% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-linear.R
\name{pca_exact}
\alias{pca_exact}
\alias{pca_irlba}
\alias{pca_random}
\title{Linear Embedding Drivers}
\usage{
pca_exact(deferred = FALSE, fold = Inf)

pca_irlba(deferred = FALSE, fold = Inf, extra.work = 7, ...)

pca_random(deferred = FALSE, fold = Inf, ...)
}
\arguments{
\item{deferred}{Logical scalar indicating whether centering/scaling should be deferred, see \code{?"\link{BiocSingular-options}"}.}

\item{fold}{Numeric scalar specifying the minimum fold-difference for cross-product calculation, see \code{?"\link{BiocSingular-options}"}.}

\item{extra.work}{Integer scalar, additional dimensionality of the workspace in \code{\link{runIrlbaSVD}}.}

\item{...}{Additional arguments to pass to \code{\link{runIrlbaSVD}} or \code{\link{runRandomSVD}}.
    This should not include any named arguments in those functions.}
}
\value{
A \code{BiocSingular::BiocSingularParam()} object.
}
\description{
Linear Embedding Drivers
}
\details{
Different engines for computing
linear embeddings (mostly PCA at this stage). Usually you
won't need to interact with these functions directly.
}
