% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TourExperiment.R
\docType{methods}
\name{neighborSets}
\alias{neighborSets}
\title{Access or set the neighborSets of a TourExperiment}
\usage{
neighborSet(x, type, ...)

neighborSet(x, type, ...) <- value

neighborSets(x, type, ...)

neighborSetNames(x, type, ...)

neighborSetNames(x, type, ...) <- value

\S4method{neighborSet}{TourExperiment,missing}(x, type)

\S4method{neighborSet}{TourExperiment,numeric}(x, type)

\S4method{neighborSet}{TourExperiment,character}(x, type)

\S4method{neighborSet}{TourExperiment,missing}(x, type, ...) <- value

\S4method{neighborSet}{TourExperiment,numeric}(x, type, ...) <- value

\S4method{neighborSet}{TourExperiment,character}(x, type, ...) <- value

\S4method{neighborSets}{TourExperiment}(x)

\S4method{neighborSetNames}{TourExperiment}(x)

\S4method{neighborSetNames}{TourExperiment}(x) <- value
}
\arguments{
\item{x}{an object (a TourExperiment)}

\item{type}{the name or a number of the list to access}

\item{value}{a matrix to assign to an element of basisSets}
}
\value{
a \code{TourExperiment} object
}
\description{
Access or set the neighborSets of a TourExperiment
}
