% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/EM-methods.R
\docType{class}
\name{NonLinearEmbeddingMatrix-class}
\alias{NonLinearEmbeddingMatrix-class}
\alias{NonLinearEmbeddingMatrix}
\title{NonLinearEmbeddingMatrix class}
\usage{
NonLinearEmbeddingMatrix(sampleFactors = matrix(nrow = 0, ncol = 0),
  featureLoadings = matrix(nrow = 0, ncol = 0), factorData = NULL,
  metadata = list(), param)
}
\arguments{
\item{sampleFactors}{A matrix-like object of sample embeddings, where rows are samples and columns are factors.}

\item{featureLoadings}{A matrix-like object of feature loadings, where rows are features and columns are factors.}

\item{factorData}{A DataFrame containing factor-level information, with one row per factor.}

\item{metadata}{An optional list of arbitrary content describing the overall experiment.}

\item{param}{An \code{NonLinearEmbeddingParam} object}
}
\description{
The NonLinearEmbeddingMatrix class is used
for storing low-dimensional embeddings from non-linear
dimensionality reduction techniques.
}
\details{
The \code{NonLinearEmbeddingMatrix} class inherits
from \code{\link[SingleCellExperiment:LinearEmbeddingMatrix]{SingleCellExperiment::LinearEmbeddingMatrix()}} and is
designed to be stored inside a \code{TourExperiment()} object. It
gains an additional slot, storing \code{NonLinearEmbeddingParam-class}
object, allowing us to maintain the arguments used to generate
the embedding matrix. This is useful for running diagnostics over
the same parameter set.
}
\seealso{
\code{\link[SingleCellExperiment:LinearEmbeddingMatrix]{SingleCellExperiment::LinearEmbeddingMatrix()}}
}
