% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors-overlay.R
\name{overlay-neighbors-diff}
\alias{overlay-neighbors-diff}
\alias{overlay_difference_neighbors}
\alias{overlay_difference_shared_neighbors}
\alias{overlay_intersect_neighbors}
\alias{overlay_intersect_shared_neighbors}
\title{Overlay a difference/intersection of two neighborSets}
\usage{
overlay_difference_neighbors(x, y, set1, set2, ...)

overlay_difference_shared_neighbors(x, y, set1, set2, type = "rank", ...)

overlay_intersect_neighbors(x, y, set1, set2, ...)

overlay_intersect_shared_neighbors(x, y, set1, set2, type = "rank", ...)
}
\arguments{
\item{x, y}{numeric vectors to produce segments from}

\item{set1, set2}{nearest neighbor graph indices to compare}

\item{type}{Algorithm used to compute shared nearest neighbors graph}
}
\description{
Overlay a difference/intersection of two neighborSets
}
