% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors-overlay.R
\name{overlay-neighbors}
\alias{overlay-neighbors}
\alias{overlay_neighbors}
\alias{overlay_shared_neighbors}
\title{Overlay a neighborSet onto a set of xy coordinates}
\usage{
overlay_neighbors(x, y, indices, ...)

overlay_shared_neighbors(x, y, indices, type = "rank", ...)
}
\arguments{
\item{x, y}{numeric vectors to produce segments from}

\item{indices}{a neighborSet index matrix}

\item{...}{additional arguments to pass to \link[ggplot2:geom_segment]{ggplot2::geom_segment}}

\item{type}{Algorithm used to compute shared nearest neighbors graph}
}
\description{
Overlay a neighborSet onto a set of xy coordinates
}
\seealso{
\link[scran:neighborsToSNNGraph]{scran::neighborsToSNNGraph}
}
