% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{overlay_kmeans_centroids}
\alias{overlay_kmeans_centroids}
\title{Overylay k-means centroids onto an xy scatter}
\usage{
overlay_kmeans_centroids(x, y, num_centers, ...)
}
\arguments{
\item{x}{numeric vectors to produce segments from}

\item{y}{numeric vectors to produce segments from}

\item{num_centers}{the number of centroids to estimate}

\item{...}{additional arguments to pass to \link[ggplot2:geom_segment]{ggplot2::geom_segment}}
}
\description{
Overylay k-means centroids onto an xy scatter
}
