% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{overlay_neighbors}
\alias{overlay_neighbors}
\title{Overlay a neighborSet on a set of xy coordinates}
\usage{
overlay_neighbors(x, y, indices, ...)
}
\arguments{
\item{x, y}{numeric vectors to produce segments from}

\item{indices}{a neighborSet index matrix}

\item{...}{additional arguments to pass to \link[ggplot2:geom_segment]{ggplot2::geom_segment}}
}
\description{
Overlay a neighborSet on a set of xy coordinates
}
