% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{overlay_shared_neighbors}
\alias{overlay_shared_neighbors}
\title{Overlay a shared nearest neighbors graph onto an xy-scatter}
\usage{
overlay_shared_neighbors(x, y, indices, type = "rank", ...)
}
\arguments{
\item{x}{numeric vectors to produce segments from}

\item{y}{numeric vectors to produce segments from}

\item{indices}{a neighborSet index matrix}

\item{type}{Algorithm used to compute shared nearest neighbors graph}

\item{...}{additional arguments to pass to \link[ggplot2:geom_segment]{ggplot2::geom_segment}}
}
\description{
Overlay a shared nearest neighbors graph onto an xy-scatter
}
