% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{overlay_snn_centroids}
\alias{overlay_snn_centroids}
\title{Overlay s-NN centroids}
\usage{
overlay_snn_centroids(x, y, indices, ...)
}
\arguments{
\item{x}{numeric vectors to produce segments from}

\item{y}{numeric vectors to produce segments from}

\item{indices}{a neighborSet index matrix}

\item{...}{additional arguments to pass to \link[ggplot2:geom_segment]{ggplot2::geom_segment}}
}
\description{
Overlay s-NN centroids
}
