% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sightse.R
\docType{class}
\name{SightSE}
\alias{SightSE}
\title{Represent a data.frame or matrix as a SightSE object.}
\usage{
SightSE(se, traveller = tourr::grand_tour())

as_sightse(.data, traveller, ...)

\S4method{as_sightse}{SingleCellExperiment,tour_path}(.data, traveller)

\S4method{as_sightse}{SummarizedExperiment,tour_path}(.data, traveller)

\S4method{as_sightse}{matrix,tour_path}(.data, traveller)

\S4method{as_sightse}{data.frame,tour_path}(.data, traveller, ...,
  view_as = "matrix")

\S4method{show}{SightSE}(object)
}
\arguments{
\item{se}{a \code{SingleCellExperiment::SingleCellExperiment()} object}

\item{traveller}{a function of class \code{tour_path} (usually tourr::grand_tour())}

\item{.data}{object to convert to a \code{SightSE} object}

\item{...}{if \code{.data} is a data.frame the columns to convert to a matrix,
every other column not included in \code{...} will become rowData}

\item{view_as}{the type of matrix if \code{.data} is a data.frame (default is numeric matrix).}
}
\description{
Represent a data.frame or matrix as a SightSE object.
}
\details{
A \code{SightSE} object inherits from the
\code{SingleCellExperiment::SingleCellExperiment()} abstraction. However, we feel
that is more general/useful than just for biological data sets.
Briefly it represents a matrix of homegenous types (n rows by p columns)
alongside \code{SummarizedExperiment::rowData()} which represents data about
the rows and \code{SummarizedExperiment::colData()} which represents data about the
columns. The \code{SightSE} object  adds an additional slot called a \code{traveller}
which is a callback function that implements a tour path, like
\code{tourr::grand_tour()}.
}
\examples{
sphere <- generate_sphere(1000, 10, mean =  5, sd = 2)
se_sphere <- as_sightse(sphere, traveller = tourr::grand_tour())
se_sphere

# convert a data.frame to a SightSE object
# this allows you to select columns that will form the assay data
se_olive <- as_sightse(tourr::olive, 
                       traveller = tourr::guided_tour(tourr::holes()),
                       palmitic:eicosenoic)
se_olive

}
