% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour.R
\name{sneezy_centroids}
\alias{sneezy_centroids}
\title{Animate a tour with a centroids overlay}
\usage{
sneezy_centroids(.data, basis, neighbor, ...)
}
\arguments{
\item{.data}{a numeric dataset}

\item{basis}{name of basisSet in .data}

\item{neighbor}{name of neighborSet in .data}

\item{...}{other control options passed to \code{\link[tourr:display_xy]{tourr::display_xy()}}}
}
\description{
Animate a tour with a centroids overlay
}
