% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour.R
\name{sneezy_neighbors}
\alias{sneezy_neighbors}
\title{Animate tour with neighborhood graph}
\usage{
sneezy_neighbors(.data, basis, neighbor, ...)
}
\arguments{
\item{.data}{a numeric dataset}

\item{basis}{name of basisSet in .data}

\item{neighbor}{name of neighborSet in .data}

\item{...}{other control options passed to \code{\link[tourr:display_xy]{tourr::display_xy()}}}

\item{.subset}{a vector of indices to find nearest neighbours, by default
all nearest neighbours will be used.}
}
\description{
Animate tour with neighborhood graph
}
