% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{sneezy_neighbours}
\alias{sneezy_neighbours}
\title{Animate grand tour with t-SNE neighbourhood graph}
\usage{
sneezy_neighbours(data, tsne_coords, .subset = NULL, ...)
}
\arguments{
\item{data}{a numeric dataset}

\item{tsne_coords}{a list from running \code{Rtsne::Rtsne()}}

\item{.subset}{a vector of indices to find nearest neighbours, by default
all nearest neighbours will be used.}

\item{...}{other control options passed to \code{\link[tourr:display_xy]{tourr::display_xy()}}}
}
\description{
Animate grand tour with t-SNE neighbourhood graph
}
