% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour.R
\name{sneezy_tour}
\alias{sneezy_tour}
\title{Animate a tour path directly with TourExperiment}
\usage{
sneezy_tour(.data, basis = 1, row_subset = NULL, apf = 1/10,
  frames = 100, height = 400, width = 400, ...)
}
\arguments{
\item{.data}{A \link{TourExperiment} object}

\item{basis}{The name or position of a \link{basisSet} in \code{.data},
defaults to the first element. The name of the \link{basisSet} will
also be used to extract the underlying data.}

\item{row_subset}{The number of rows to subset the view on, the default
which is NULL will select all rows.}

\item{apf}{angles per frame (defaults to 1/30)}

\item{frames}{number of frames in animation (defaults to 300)}

\item{height, width}{the size of the animation in pixels}

\item{...}{other control options passed to \code{\link[tourr:display_xy]{tourr::display_xy()}}}
}
\value{
A \link[gganimate:gif_file]{gganimate::gif_file} that can be embedded in an Rmd or
R console viwer pane.
}
\description{
Animate a tour path directly with TourExperiment
}
\examples{
multi_te <- TourExperiment(multi, X1:X10)
multi_te <- generate_bases(multi_te, .on = "view")
multi_te <- estimate_neighbors(multi_te, 10, .on = "view")

sneezy_neighbors(multi_te)

}
\seealso{
\code{\link[tourr:animate_xy]{tourr::animate_xy()}}, \code{\link[tourr:display_xy]{tourr::display_xy()}}
}
