% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{view_tour_xy}
\alias{view_tour_xy}
\title{Project a data set onto a basis set}
\usage{
view_tour_xy(.data, .on = NULL, .subset = NULL, .color = NULL,
  clamp = TRUE, aps = 1, fps = NULL, ...)
}
\arguments{
\item{.data}{A \code{TourExperiment} object}

\item{.on}{the name of the basisSet to tour (default = NULL, will be first basisSet)}

\item{.subset}{which points to select?}

\item{.color}{Colour points by a given variable}

\item{clamp}{Clamp all variables to lie in \link{0,1} (default = FALSE)}

\item{aps}{Target angular velocity (how quickly to move between bases?) In radians per second. Defaults to 1.}

\item{fps}{Target frame rate for streaming (defaults to 30)}

\item{...}{Other arguments}
}
\description{
Project a data set onto a basis set
}
