% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\docType{methods}
\name{view_xy,TourExperiment-method}
\alias{view_xy,TourExperiment-method}
\title{Simple interactive xy-plots for TourExperiment objects}
\usage{
\S4method{view_xy}{TourExperiment}(.data, .on = NULL, .x, .y, .color,
  ...)
}
\arguments{
\item{.data}{A \code{TourExperiment} object}

\item{.on}{Where to scatter? The assay name or reducedDim slot name}

\item{.x}{The x aesthetic (a character or integer)}

\item{.y}{The y aesthetic (a character or integer)}

\item{.color}{The colour aesthetic (either a vector or dimensions from \code{.on})}
}
\description{
Simple interactive xy-plots for TourExperiment objects
}
