% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{view_xy}
\alias{view_xy}
\alias{view_xy,DFrame-method}
\alias{view_xy,LinearEmbeddingMatrix-method}
\alias{view_xy,matrix-method}
\title{Create a simple interactive scatter-plot}
\usage{
view_xy(.data, .on = NULL, .x, .y, .color, ...)
}
\arguments{
\item{.data}{A \code{TourExperiment} object}

\item{.on}{which part of \code{.data} to select}

\item{.x}{the x-aesthetic}

\item{.y}{the y-aesthetic}

\item{.color}{the the color-aesthetic}

\item{...}{additional args}
}
\description{
Create a simple interactive scatter-plot
}
