% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{basic_tsne}
\alias{basic_tsne}
\title{Simplified running of t-SNE}
\usage{
basic_tsne(data, perplexity)
}
\arguments{
\item{data}{a data.frame or matrix to compute t-SNE on}

\item{perplexity}{the perpelxity paramater to t-SNE algorithm (see \code{Rtsne::Rtsne()})}
}
\description{
Simplified running of t-SNE
}
\details{
This a very minimal wrapper to \code{Rtsne::Rtsne()}
that computes exact t-SNE for a given dataset, withouth
normalising the input data or running PCA beforehand.
}
