% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/embed-linear.R
\docType{methods}
\name{embed_linear}
\alias{embed_linear}
\alias{embed_linear,BiocSingularParam-method}
\alias{embed_linear,missing-method}
\title{Compute a linear embedding over a \code{TourExperiment} object.}
\usage{
embed_linear(.data, num_comp, .on = NULL, center = TRUE,
  scale = FALSE, .subset = NULL,
  .parallel = BiocParallel::SerialParam(), .engine)

\S4method{embed_linear}{missing}(.data, num_comp, .on = NULL,
  center = TRUE, scale = FALSE, .subset = NULL,
  .parallel = BiocParallel::SerialParam(), .engine)

\S4method{embed_linear}{BiocSingularParam}(.data, num_comp, .on = NULL,
  center = TRUE, scale = FALSE, .subset = NULL,
  .parallel = BiocParallel::SerialParam(), .engine)
}
\arguments{
\item{.data}{A \code{TourExperiment()} object}

\item{num_comp}{Number of components to retain}

\item{.on}{The named element in \code{.data} to compute the PCA (default is NULL which
is the first assay in \code{.data}.}

\item{center}{Should columns be centered? Default = TRUE}

\item{scale}{Should columns be scaled to unit variance?}

\item{.subset}{Restrict linear embedding to run on a subset of rows. (Default = NULL).}

\item{.parallel}{A `BiocParallel::BPPARAM()`` object, default is to compute in serial.}

\item{.engine}{How to compute the embedding. If missing, defaults to \code{pca_exact()}.}
}
\description{
Compute a linear embedding over a \code{TourExperiment} object.
}
\details{
This function is a wrapper to \code{BiocSingular::runPCA()}, with
additions for computing on parts of \code{TourExperiment} object. This function
always returns a \code{TourExperiment} with the \code{reducedDim} slot updated
with \code{SingleCellExperiment::LinearEmbeddingMatrix} containing the
sample factors, loadings, and factor data from the principal components.
}
