% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-nonlinear.R
\name{embed_nonlinear}
\alias{embed_nonlinear}
\title{Compute a non-linear embedding on a \code{SightSE} object.}
\usage{
embed_nonlinear(.data, num_comp, .on, center = TRUE, scale = FALSE,
  .parallel = BiocParallel::SerialParam(), .engine)
}
\arguments{
\item{.data}{A \code{sneezy::SightSE()} object}

\item{num_comp}{Number of components/dimensions to retain}

\item{.on}{Where to compute the non-linear embedding?}

\item{center}{Should columns be centered? Default = TRUE}

\item{scale}{Should  columns be scaled to unit variance?}

\item{.parallel}{A BiocParallel::BPPARAM() object}

\item{.engine}{A BiocSingular::BiocSingular() object
used to compute the embedding. Deafults to \code{tsne_exact()}}
}
\description{
Compute a non-linear embedding on a \code{SightSE} object.
}
