% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{estimate_neighbors}
\alias{estimate_neighbors}
\title{Esiimate k-nearest neighbours graph from a TourExperiment object}
\usage{
estimate_neighbors(.data, num_neighbors, from = NULL, .engine)
}
\arguments{
\item{.data}{a \code{TourExperiment} object}

\item{num_neighbors}{An integer scalar for number of neighbors}

\item{from}{a character scalar, indicating the part of \code{.data} to
estimate nearest neighbors. If missing the first assay will be used.}

\item{.engine}{A \code{BiocNeighbor::BiocNeighborParam()} object that
reperesents the algorithm used to compute nearest neighbors.}
}
\value{
A \code{TourExperiment} object with the \code{neighborSets} slot
filled.
}
\description{
Esiimate k-nearest neighbours graph from a TourExperiment object
}
\details{
The representation of the nearest neighbors is a list with
two elements, one called \code{index} containing a matrix of the nearest
neighbor indexes and another \code{distance} containing the distance
to each neighbor.
}
