% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{get_neighbourhood_graph}
\alias{get_neighbourhood_graph}
\title{Generate nearest neighbours graph in t-SNE space}
\usage{
get_neighbourhood_graph(tsne_coords)
}
\arguments{
\item{tnse_coords}{the list obtained from running \code{Rtsne::Rtsne()}}
}
\description{
Generate nearest neighbours graph in t-SNE space
}
