% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-linear.R
\name{pca_exact}
\alias{pca_exact}
\alias{pca_irlba}
\alias{pca_random}
\title{Linear Embedding Drivers}
\usage{
pca_exact(deferred = FALSE, fold = Inf)

pca_irlba(deferred = FALSE, fold = Inf, extra.work = 7, ...)

pca_random(deferred = FALSE, fold = Inf, ...)
}
\value{
A \code{BiocSingular::BiocSingularParam()} object.
}
\description{
Linear Embedding Drivers
}
\details{
Different engines for computing
linear embeddings (mostly PCA at this stage). Usually you
won't need to interact with these functions directly.
}
