% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids-overlay.R
\name{overlay-centroids}
\alias{overlay-centroids}
\alias{overlay_knn_centroids}
\alias{overlay_snn_centroids}
\alias{overlay_kmeans_centroids}
\title{Overlay centroids via k-means or graph-based clustering}
\usage{
overlay_knn_centroids(x, y, indices, ...)

overlay_snn_centroids(x, y, indices, ...)

overlay_kmeans_centroids(x, y, num_centers, ...)
}
\arguments{
\item{num_centers}{the number of centroids to estimate}
}
\value{
a \link[ggplot2:geom_point]{ggplot2::geom_point} layer
}
\description{
Overlay centroids via k-means or graph-based clustering

Overylay k-means centroids onto an xy scatter
}
\details{
The nearest neighbor centroids functions
automatically estimate centroids from an indices matrix
representing a nearest neighbor graph via a community detection
algorithm (louvain) from the \link{igraph} package.
The smaller the value of \code{ncol(indices)} the more centroids
there will be. The kmeans method returns the centroids from
run \link[stats:kmeans]{stats::kmeans} with the user suppplied number of centers.
}
