% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdfsense}
\alias{pdfsense}
\title{pdfsense data}
\format{An object of class \code{TourExperiment} with 56 rows and 2808 columns.}
\source{
\url{http://www.physics.smu.edu/botingw/PDFsense_web_histlogy}
}
\usage{
pdfsense
}
\description{
pdfsense data
}
\details{
56 variables summarised over many particle physics experiments
}
\keyword{datasets}
