% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/TE-constructor.R
\docType{class}
\name{TourExperiment-class}
\alias{TourExperiment-class}
\alias{TourExperiment}
\alias{show,TourExperiment-method}
\alias{TourExperiment,SingleCellExperiment-method}
\alias{TourExperiment,SummarizedExperiment-method}
\alias{TourExperiment,matrix-method}
\alias{TourExperiment,data.frame-method}
\title{Represent a data.frame or matrix or assay-like object as a \code{TourExperiment}}
\usage{
TourExperiment(.data, ..., basisSets = SimpleList(),
  neighborSets = SimpleList())

\S4method{show}{TourExperiment}(object)

\S4method{TourExperiment}{SingleCellExperiment}(.data,
  basisSets = SimpleList(), neighborSets = SimpleList())

\S4method{TourExperiment}{SummarizedExperiment}(.data,
  basisSets = SimpleList(), neighborSets = SimpleList())

\S4method{TourExperiment}{matrix}(.data, ..., assayName = "view",
  basisSets = SimpleList(), neighborSets = SimpleList())

\S4method{TourExperiment}{data.frame}(.data, ..., assayName = "view",
  assayType = "matrix", basisSets = SimpleList(),
  neighborSets = SimpleList())
}
\arguments{
\item{.data}{object to convert to a \code{TourExperiment} object}

\item{...}{if \code{.data} is a data.frame the columns to convert to a matrix,
every other column not included in \code{...} will become colData, the names
of the columns selected with \code{...} will be become row names in the resulting
object.}

\item{basisSets}{a SimpleList object containing tour bases. Default is
an empty list.}

\item{neighborSets}{a SimpleList conatining nearest neighbours Default
is an empty list.}

\item{object}{a \code{TourExperiment} object}

\item{assayName}{the character name of the \code{assay} slot in the \code{TourExperiment} object.
Defaults to 'view'.}

\item{assayType}{the type of matrix if \code{.data} is a data.frame. Defaults to
'matrix'.}
}
\description{
Represent a data.frame or matrix or assay-like object as a \code{TourExperiment}
}
\details{
A \code{TourExperiment} object inherits from the
\code{SingleCellExperiment::SingleCellExperiment()} class. However, we feel
that is more general/useful than just for biological data sets.
Briefly, it represents a matrix of homegenous types (n rows 'features' by
p columns 'samples'). Each 'feature' is aligned to
\code{SummarizedExperiment::rowData()} slot which represents data about
the rows and \code{SummarizedExperiment::colData()} which represents data about the
samples. The \code{TourExperiment} object  includes two additional slots. The first
is a slot called \code{basisSets} is a \code{SimpleList}, which represents the
anchor bases from computing a tour. The second is a slot called \code{neighborSets}
which is a \code{SimpleList} that represents nearest
neighbour indexes and distances.
}
\examples{
sphere <- generate_sphere(1000, 10, mean =  5, sd = 2)
te_sphere <- TourExperiment(sphere)
te_sphere

# convert a data.frame to a TourExperiment object
# this allows you to select columns that will form the assay data
multi_te <- TourExperiment(multi, X1:X10)
multi_te

}
\seealso{
\code{\link[=generate_bases]{generate_bases()}}, \code{\link[=basisSets]{basisSets()}},
\code{\link[=estimate_neighbors]{estimate_neighbors()}}, \code{\link[=neighborSets]{neighborSets()}}
}
