% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TourExperiment.R
\docType{class}
\name{TourExperiment}
\alias{TourExperiment}
\title{Represent a data.frame or matrix as a `TourExperiment``}
\usage{
TourExperiment(.data, basisSets = S4Vectors::SimpleList(),
  neighborSets = S4Vectors::SimpleList(), ...)

\S4method{TourExperiment}{SingleCellExperiment}(.data,
  basisSets = S4Vectors::SimpleList(),
  neighborSets = S4Vectors::SimpleList(), ...)

\S4method{TourExperiment}{SummarizedExperiment}(.data,
  basisSets = S4Vectors::SimpleList(),
  neighborSets = S4Vectors::SimpleList(), ...)

\S4method{TourExperiment}{matrix}(.data,
  basisSets = S4Vectors::SimpleList(),
  neighborSets = S4Vectors::SimpleList(), ...)

\S4method{TourExperiment}{data.frame}(.data, basisSets, neighborSets, ...,
  viewAs = "matrix")

\S4method{show}{TourExperiment}(object)
}
\arguments{
\item{.data}{object to convert to a \code{TourExperiment} object}

\item{basisSets}{a SimpleList object containing tour bases}

\item{neighborSets}{a SimpleList conatining nearest neighbours}

\item{...}{if \code{.data} is a data.frame the columns to convert to a matrix,
every other column not included in \code{...} will become rowData}

\item{viewAs}{the type of matrix if \code{.data} is a data.frame (default is numeric matrix).}
}
\description{
Represent a data.frame or matrix as a `TourExperiment``
}
\details{
A \code{TourExperiment} object inherits from the
\code{SingleCellExperiment::SingleCellExperiment()} class. However, we feel
that is more general/useful than just for biological data sets.
Briefly, it represents a matrix of homegenous types (n rows by p columns)
alongside \code{SummarizedExperiment::rowData()} which represents data about
the rows and \code{SummarizedExperiment::colData()} which represents data about the
columns. The \code{TourExperiment} object  adds two additional slots. The first
is a slot called \code{basisSets} is a \code{SimpleList}, which represents the
anchor bases from computing a tour. The second is a slot called \code{neighborSets}
which is a \code{SimpleList} that represents nearest
neighbour indexes and distances.
}
\examples{
sphere <- generate_sphere(1000, 10, mean =  5, sd = 2)
te_sphere <- TourExperiment(sphere)
te_sphere

# convert a data.frame to a TourExperiment object
# this allows you to select columns that will form the assay data
te_olive <- TourExperiment(tourr::olive, palmitic:eicosenoic)
te_olive

}
